+++ client/src/component_ui.rs
@@ -1,0 +1,2 @@
+pub struct ComponentUi {
+}
